FF: Move masthead to Header namespace.
