+++ wp-content/plugins/core/src/Templates/Controllers/Header/Masthead.php
@@ -1,7 +1,7 @@
 <?php
 declare( strict_types=1 );
 
-namespace Tribe\Project\Templates\Controllers\Document;
+namespace Tribe\Project\Templates\Controllers\Header;
 
 use Tribe\Project\Templates\Abstract_Controller;
 use Tribe\Project\Templates\Component_Factory;
